poll_hyperion_ng_tag.yml: disable for now
This workflow is broken due to changes in how GitHub token permissions
work. It seems a PAT is now needed for a workflow to make changes to
workflow files. I've tried adding a permission to see if that helps,
though.

I also stopped the workflow from running automatically until this issue
is resolved.